# src/swagger/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Pod Quiz Documentation",
    "description": "API documentation for Pod Project",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.2"
  },
  "servers": [
    {
      "url": "http://localhost:9200"
    },
    {
      "url": "http://172.20.10.12:9200"
    }
  ],
  "paths": {
    "/api/account/register": {
      "post": {
        "tags": [
          "Web-Account"
        ],
        "summary": "Register a new account",
        "operationId": "registerAccount",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phoneNumber": {
                    "type": "string",
                    "example": "+1234567890"
                  },
                  "firstName": {
                    "type": "string",
                    "example": "John"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "Doe"
                  },
                  "password": {
                    "type": "string",
                    "example": "Password123!"
                  },
                  "role": {
                    "type": "string",
                    "example": "user"
                  }
                },
                "required": [
                  "phoneNumber",
                  "firstName",
                  "lastName",
                  "password",
                  "role"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account successfully registered",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "200"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "server internal error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/account/verify": {
      "post": {
        "tags": [
          "Web-Account"
        ],
        "summary": "Verify an account",
        "operationId": "verifyAccount",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "otp": {
                    "type": "string",
                    "example": "12345"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "example": "+1234567890"
                  }
                },
                "required": [
                  "otp",
                  "phoneNumber"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account successfully verified",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "registered successfully"
                    },
                    "token": {
                      "type": "string",
                      "example": "accessToken12345"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "OTP not verified",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "otp not verified"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "OTP or user not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "server internal error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/account/login": {
      "post": {
        "tags": [
          "Web-Account"
        ],
        "summary": "Login to an account",
        "operationId": "loginByPass",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phoneNumber": {
                    "type": "string",
                    "example": "+1234567890"
                  },
                  "password": {
                    "type": "string",
                    "example": "Password123!"
                  }
                },
                "required": [
                  "phoneNumber",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "60d0fe4f5311236168a109ca"
                        },
                        "phoneNumber": {
                          "type": "string",
                          "example": "+1234567890"
                        },
                        "firstName": {
                          "type": "string",
                          "example": "John"
                        },
                        "lastName": {
                          "type": "string",
                          "example": "Doe"
                        },
                        "role": {
                          "type": "string",
                          "example": "user"
                        },
                        "isVerify": {
                          "type": "boolean",
                          "example": true
                        },
                        "accessToken": {
                          "type": "string",
                          "example": "accessToken12345"
                        },
                        "refreshToken": {
                          "type": "string",
                          "example": "refreshToken12345"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Wrong credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "wrong data"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "user not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "server internal error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/cart": {
      "post": {
        "tags": ["Web-Cart"],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Add or update cart",
        "operationId": "newCart",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "order": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "restaurant": {
                          "type": "string",
                          "example": "60d0fe4f5311236168a109cf"
                        },
                        "item": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "menuItem": {
                                "type": "string",
                                "example": "60d0fe4f5311236168a109d0"
                              }
                            },
                            "required": [
                              "menuItem"]
                          }
                        }
                      },
                      "required": [
                        "restaurant",
                        "item"
                      ]
                    }
                  }
                },
                "required": [
                  "order"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cart updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Cart updated successfully"
                    },
                    "cart": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "60d0fe4f5311236168a109d1"
                        },
                        "user": {
                          "type": "string",
                          "example": "60d0fe4f5311236168a109cb"
                        },
                        "order": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "restaurant": {
                                "type": "string",
                                "example": "60d0fe4f5311236168a109cf"
                              },
                              "item": {
                                "type": "array",
                                "items": {
                                  "type": "object",
                                  "properties": {
                                    "menuItem": {
                                      "type": "string",
                                      "example": "60d0fe4f5311236168a109d0"
                                    },
                                    "quantity": {
                                      "type": "integer",
                                      "example": 2
                                    }
                                  }
                                }
                              }
                            }
                          }
                        },
                        "totalAmount": {
                          "type": "number",
                          "example": 29.99
                        },
                        "status": {
                          "type": "string",
                          "example": "pending"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2024-07-22T14:48:00.000Z"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2024-07-22T14:48:00.000Z"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Server internal error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
...
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "name": "x-access-token",
        "in": "header"
      }
    }
  }
}